pyproject.toml
[project]
name = "russian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "google-genai>=1.52.0",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.51.0",
    "streamlit-mic-recorder>=0.0.8",
]
